+++ htf/utils.py
@@ -65,9 +65,10 @@ def compute_ohe_bead_type_interactions(pos_btype, nlist_btype, n_btypes):
     '''
     m, n = tf.math.minimum(pos_btype[..., tf.newaxis], nlist_btype), tf.math.maximum(
         pos_btype[..., tf.newaxis], nlist_btype)
-    one_hot_indices = m*(2*n_btypes - m + 1)//2 + n - m
-    # Finding the total number of possible interactions between different bead types
-    total_interactions = n_btypes * (n_btypes-1) // 2 + n_btypes
+    one_hot_indices = m * (2 * n_btypes - m + 1) // 2 + n - m
+    # Finding the total number of possible interactions between different bead
+    # types
+    total_interactions = n_btypes * (n_btypes - 1) // 2 + n_btypes
     return tf.one_hot(one_hot_indices, depth=total_interactions)
 
 
@@ -147,11 +148,11 @@ def compute_nlist(
 
     if return_types:
         nlist_type = tf.reshape(
-            tf.gather(positions[:, 3], flat_idx[:, 0]), [-1, NN, 1])
+            tf.gather(positions[:, 3], flat_idx[:, 1]), [-1, NN, 1])
         return tf.concat([
             nlist_pos,
-            nlist_type * nlist_mask
-        ], axis=-1)
+            nlist_type
+        ], axis=-1) * nlist_mask
     else:
         return tf.concat([
             nlist_pos,
@@ -352,7 +353,11 @@ def find_cgnode_id(atm_id, cg):
                 return num_index
 
 
-def gen_mapped_exclusion_list(universe, atoms_in_molecule, beads_mappings, selection='all'):
+def gen_mapped_exclusion_list(
+        universe,
+        atoms_in_molecule,
+        beads_mappings,
+        selection='all'):
     ''' Generates mapped exclusion list to compute mapped_nlist for non-bonded bead-type
     interactions.
 
@@ -367,7 +372,8 @@ def gen_mapped_exclusion_list(universe, atoms_in_molecule, beads_mappings, selec
     :type selection: string
 
     :return: A [B x B] array of dtype bools, indicating which pairs
-        should be excluded from nlist (True = exclude)
+        should be excluded from nlist (True = exclude). B is the total number of
+        bead particles in the system
     '''
     # Get the number of atoms
     N = len(universe.select_atoms(selection))
@@ -379,7 +385,7 @@ def gen_mapped_exclusion_list(universe, atoms_in_molecule, beads_mappings, selec
     matrix_mapping_molecule = hoomd.htf.matrix_mapping(
         atoms_in_molecule, beads_mappings, mass_weighted=False)[1]
     # Get the number of molecules
-    M = N//matrix_mapping_molecule.shape[1]
+    M = N // matrix_mapping_molecule.shape[1]
     # repeat matrix_mapping_molecule along diag
     matrix_mapping_system = np.kron(
         np.eye(M, dtype=int), matrix_mapping_molecule).astype(bool)
@@ -387,6 +393,22 @@ def gen_mapped_exclusion_list(universe, atoms_in_molecule, beads_mappings, selec
         matrix_mapping_system.T)
     np.fill_diagonal(mapped_exclusion, False)
     return mapped_exclusion
+
+
+def gen_bonds_group(mapped_exclusion_list):
+    ''' Generates HOOMD snapshot bonds group based on
+    bead particles' exclusion list.
+
+    :param mapped_exclusion_list: A [B x B] array of dtype bools,
+        indicating which pairs should be excluded from nlist (True = exclude)
+    :type obj: numpy array
+
+    :return: :py:meth:`hoomd.data.make_snapshot.bonds.group`
+    '''
+    rows, cols = np.where(mapped_exclusion_list)
+    bonds_group = np.array([[rows[i], cols[i]]
+                            for i in range(rows.shape[0]) if rows[i] <= cols[i]])
+    return bonds_group
 
 
 def compute_adj_mat(obj):
@@ -566,6 +588,48 @@ def compute_cg_graph(
             return rs, angs, dihs
 
 
+def mol_features_multiple(bnd_indices=None, ang_indices=None,
+                          dih_indices=None, molecules=None, beads=None):
+    '''Use to apply the result of compute_cg_graph to multiple
+    molecules in a system. Computes the atom/cg bead indices of bonds,angles
+    and dihedrals for multiple molecules.
+
+    :param bnd_indices: indices of atoms making bonds
+    :type bnd_indices: numpy array of type int
+    :param ang_indices: indices of atoms making angles
+    :type ang_indices: numpy array of type int
+    :param dih_indices: indices of atoms making dihedrals
+    :type dih_indices: numpy array of type int
+    :param molecules: number of molecules in the system
+    :type molecules: int
+    :param beads: number of atoms/CG beads in a molecule
+    :type beads: int
+
+    :return: arrays of indices of atoms/cg beads making bonds,
+    angles and dihedrals in a system of molecules.
+    '''
+
+    bnd_ids = []
+    ang_ids = []
+    dih_ids = []
+
+    for n in range(molecules):
+        if bnd_indices is not None:
+            bnd_ids.append(bnd_indices + n * beads)
+
+        if ang_indices is not None:
+            ang_ids.append(ang_indices + n * beads)
+
+        if dih_indices is not None:
+            dih_ids.append(dih_indices + n * beads)
+
+    bnd_ids = np.asarray(bnd_ids).reshape((-1, 2))
+    ang_ids = np.asarray(ang_ids).reshape((-1, 3))
+    dih_ids = np.asarray(dih_ids).reshape((-1, 4))
+
+    return bnd_ids, ang_ids, dih_ids
+
+
 def iter_from_trajectory(
         nneighbor_cutoff,
         universe,
@@ -597,9 +661,9 @@ def iter_from_trajectory(
     :type r_cut: float
     :param period: Period of reading the trajectory frames
     :type period: int
-    :param start: Start time (ns) of reading the trajectory frames
+    :param start: Starting frame for reading the trajectory frames
     :type period: int
-    :param period: End time (ns) reading the trajectory frames
+    :param period: Ending frame for reading the trajectory frames
     :type period: int
     '''
     import MDAnalysis
@@ -680,9 +744,10 @@ def iter_from_trajectory(
             values=tf.reshape(box, (-1,)),
             dense_shape=(len(atom_group), 3, 3)
         )
-    # Run the model at every nth frame where time is in range [start,end] and n = period
+    # Run the model at every nth frame where time is in range [start,end] and
+    # n = period
     for i, ts in enumerate(tqdm(universe.trajectory)):
-        if ts.time >= start and ts.time <= end:
+        if ts.frame >= start and ts.frame <= end:
             if i % period == 0:
                 yield [nlist, np.concatenate(
                     (atom_group.positions,
@@ -736,7 +801,8 @@ def mol_angle(
         cg_positions=None,
         b1=None,
         b2=None,
-        b3=None):
+        b3=None,
+        box=None):
     ''' This method calculates the bond angle given three atoms batched by molecule.
     Or to output CG angles input CG=True and indices of the CG beads making the angles.
     cg_positions and bead indices can be computed by calling :py:meth:`.generate_cg_graph()`
@@ -760,6 +826,8 @@ def mol_angle(
     :type b2: int
     :param b3: index of third CG bead
     :type b3: int
+    :param box: low, high coordinates and tilt vectors of the box
+    :type box: [3,3] array
 
     :returns: angles:Tensor containing all atom angles (CG=False)
               or
@@ -771,24 +839,28 @@ def mol_angle(
     if mol_positions is not None and CG is False:
         v_ij = mol_positions[:, type_i, :3] - mol_positions[:, type_j, :3]
         v_jk = mol_positions[:, type_k, :3] - mol_positions[:, type_j, :3]
-        cos_a = tf.einsum('ij,ij->i', v_ij, v_jk)
+        wrap_vij = hoomd.htf.wrap_vector(v_ij, box)
+        wrap_vjk = hoomd.htf.wrap_vector(v_jk, box)
+        cos_a = tf.einsum('ij,ij->i', wrap_vij, wrap_vjk)
         cos_a = tf.math.divide(
             cos_a,
             (tf.norm(
-                tensor=v_ij,
+                tensor=wrap_vij,
                 axis=1) *
                 tf.norm(
-                tensor=v_jk,
+                tensor=wrap_vjk,
                 axis=1)))
         angles = tf.math.acos(cos_a)
         return angles
 
-    if CG and cg_positions is None:
+    if CG is True and cg_positions is None:
         raise ValueError('cg_positions not found.')
 
-    if CG and cg_positions is not None:
+    if CG is True and cg_positions is not None:
         v_ij = cg_positions[b2] - cg_positions[b1]
         v_jk = cg_positions[b3] - cg_positions[b2]
+        wrap_vij = hoomd.htf.wrap_vector(v_ij, box)
+        wrap_vjk = hoomd.htf.wrap_vector(v_jk, box)
         if type(cg_positions) == tf.Tensor:
             cos_a = tf.reduce_sum(v_ij * tf.transpose(v_jk))
             cos_a = tf.divide(cos_a, tf.norm(v_ij) * tf.norm(v_jk))
@@ -797,6 +869,12 @@ def mol_angle(
             cos_a = np.dot(v_ij, v_jk)
             cos_a = np.divide(cos_a, (np.linalg.norm(v_ij) * np.linalg.norm(v_jk)))
             cg_angles = np.arccos(cos_a)
+
+        cos_a = np.dot(wrap_vij, wrap_vjk)
+        cos_a = np.divide(
+            cos_a, (np.linalg.norm(wrap_vij) * np.linalg.norm(wrap_vjk)))
+
+        cg_angles = np.arccos(cos_a)
         return cg_angles
 
 
@@ -807,7 +885,8 @@ def mol_bond_distance(
         CG=False,
         cg_positions=None,
         b1=None,
-        b2=None):
+        b2=None,
+        box=None):
     ''' This method calculates the bond distance given two atoms batched by molecule.
     Or to output CG bond distances, input CG=True and indices of the CG bead pairs
     cg_positions and bead indices can be computed by calling :py:meth:`.generate_cg_graph()`
@@ -827,7 +906,8 @@ def mol_bond_distance(
     :type b1: int
     :param b2: index of second CG bead
     :type b2: int
-
+    :param box: low, high coordinates and tilt vectors of the box
+    :type box: [3,3] array
 
     :returns: v_ij: Tensor containing bond distances(CG=False)
               or
@@ -839,19 +919,21 @@ def mol_bond_distance(
 
     if CG is False and mol_positions is not None:
         v_ij = mol_positions[:, type_j, :3] - mol_positions[:, type_i, :3]
-        v_ij = tf.norm(tensor=v_ij, axis=1)
-        return v_ij
+        wrap_vij = hoomd.htf.wrap_vector(v_ij, box)
+        wrap_vij = tf.norm(tensor=wrap_vij, axis=1)
+        return wrap_vij
 
-    if CG and cg_positions is None:
+    if CG is True and cg_positions is None:
         raise ValueError('cg_positions not found')
 
-    if CG and cg_positions is not None:
+    if CG is True and cg_positions is not None:
         u_ij = cg_positions[b2] - cg_positions[b1]
+        wrap_uij = hoomd.htf.wrap_vector(u_ij, box)
         if type(cg_positions) == tf.Tensor:
-            u_ij = tf.norm(u_ij)
+            wrap_uij = tf.norm(wrap_uij)
         else:
-            u_ij = np.linalg.norm(u_ij)
-        return u_ij
+            wrap_uij = np.linalg.norm(wrap_uij)
+        return wrap_uij
 
 
 def mol_dihedral(
@@ -865,7 +947,8 @@ def mol_dihedral(
         b1=None,
         b2=None,
         b3=None,
-        b4=None):
+        b4=None,
+        box=None):
     ''' This method calculates the dihedral angles given three atoms batched by molecule.
     Or to output CG dihedral angles input CG=True and indices of the CG beads making the angles.
     cg_positions and bead indices can be computed by calling :py:meth:`.generate_cg_graph()`
@@ -893,6 +976,8 @@ def mol_dihedral(
     :type b3: int
     :param b4: index of fourth CG bead
     :type b4: int
+    :param box: low, high coordinates and tilt vectors of the box
+    :type box: [3,3] array
 
     :returns: dihedrals:Tensor containing all atom dihedral angles (CG=False)
               or
@@ -906,10 +991,13 @@ def mol_dihedral(
         v_ij = mol_positions[:, type_j, :3] - mol_positions[:, type_i, :3]
         v_jk = mol_positions[:, type_k, :3] - mol_positions[:, type_j, :3]
         v_kl = mol_positions[:, type_l, :3] - mol_positions[:, type_k, :3]
+        wrap_vij = hoomd.htf.wrap_vector(v_ij, box)
+        wrap_vjk = hoomd.htf.wrap_vector(v_jk, box)
+        wrap_vkl = hoomd.htf.wrap_vector(v_kl, box)
 
         # calculation of normal vectors
-        n1 = tf.linalg.cross(v_ij, v_jk)
-        n2 = tf.linalg.cross(v_jk, v_kl)
+        n1 = tf.linalg.cross(wrap_vij, wrap_vjk)
+        n2 = tf.linalg.cross(wrap_vjk, wrap_vkl)
         n1_norm = tf.norm(tensor=n1)
         n2_norm = tf.norm(tensor=n2)
         if n1_norm == 0.0 or n2_norm == 0.0:
@@ -921,13 +1009,16 @@ def mol_dihedral(
         dihedrals = tf.math.acos(cos_d)
         return dihedrals
 
-    if CG and cg_positions is None:
+    if CG is True and cg_positions is None:
         raise ValueError('cg_positions not found.')
 
-    if CG and cg_positions is not None:
+    if CG is True and cg_positions is not None:
         v_ij = cg_positions[b2] - cg_positions[b1]
         v_jk = cg_positions[b3] - cg_positions[b2]
         v_kl = cg_positions[b4] - cg_positions[b3]
+        wrap_vij = hoomd.htf.wrap_vector(v_ij, box)
+        wrap_vjk = hoomd.htf.wrap_vector(v_jk, box)
+        wrap_vkl = hoomd.htf.wrap_vector(v_kl, box)
 
         if type(cg_positions) == tf.Tensor:
             n1 = tf.linalg.cross(v_ij, v_jk)